+++ CHANGELOG.md
@@ -13,6 +13,7 @@ You can also check the
 
 - Fixes
   - Text of locale switcher select options is now visible on Windows machines
+  - Total value labels do not overlap with error bars anymore
 - Styles
   - Header of data preview table is now sticky
   - Data preview table should now be aligned with the design